Move bindings to Python module
